requirements.txt
gTTS
playsound
pillow
pytesseract
opencv-python
PyQt5
keyboard
